requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-doc
annotated-types
anyio
attrs
backoff
bcrypt
build
certifi
cffi
charset-normalizer
chromadb
click
clickhouse-connect
cryptography
dataclasses-json
distro
dnspython
duckdb
durationpy
email-validator
fastapi
fastapi-cli
fastapi-cloud-cli
fastar
filelock
filetype
flatbuffers
frozenlist
fsspec
ftfy
google-auth
google-genai
googleapis-common-protos
grpcio
h11
hf-xet
hnswlib
httpcore
httptools
httpx
httpx-sse
huggingface_hub
idna
importlib_metadata
importlib_resources
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-classic
langchain-community
langchain-core
langchain-google-genai
langchain-ollama
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lz4
markdown-it-py
MarkupSafe
marshmallow
mdurl
mmh3
mpmath
multidict
mypy_extensions
networkx
numpy
oauthlib
ollama
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
ormsgpack
overrides
packaging
pandas
pillow
posthog
propcache
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
pypdf
PyPika
pyproject_hooks
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rich-toolkit
rignore
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
sentry-sdk
shellingham
six
sniffio
SQLAlchemy
starlette
sympy
tenacity
threadpoolctl
tokenizers
torch
tqdm
transformers
typer
typer-slim
typhoon-ocr
typing-inspect
typing-inspection
typing_extensions
urllib3
uuid_utils
uvicorn
uvloop
watchfiles
wcwidth
websocket-client
websockets
xxhash
yarl
zipp
zstandard